Move reservations-related code to reservation.py
