+++ src/xpk/core/capacity.py
@@ -15,12 +15,15 @@ limitations under the License.
 """
 
 import enum
-from dataclasses import dataclass
 
-from .commands import run_command_with_updates, run_command_for_value
 from .system_characteristics import AcceleratorType
-from ..utils.console import xpk_print, xpk_exit
+from ..utils.console import xpk_print
 from ..utils.kueue import is_queued_cluster
+from .reservation import (
+    ReservationLink,
+    verify_reservations_exist,
+    to_reservation_path,
+)
 
 AUTOPROVISIONING_CONFIG_VALUE = 'AUTOPROVISION'
 AUTOPROVISIONING_CONFIG_MINIMUM_KEY = 'minimum_chips'
@@ -33,7 +36,6 @@ H200_DEVICE_TYPE = 'h200-141gb-8'
 B200_DEVICE_TYPE = 'b200-8'
 GB200_DEVICE_TYPE = 'gb200-4'
 GB200_DEVICE_TYPE_NOLSSD = 'gb200-4-no-ssd'
-RESERVATION_CONFIG_KEY = 'reservation_id'
 
 
 class CapacityType(enum.Enum):
@@ -42,42 +44,6 @@ class CapacityType(enum.Enum):
   SPOT = 'spot'
   UNKNOWN = 'unknown'
   FLEX_START = 'flex_start'
-
-
-@dataclass(frozen=True)
-class ReservationLink:
-  project: str
-  name: str
-  zone: str
-
-
-@dataclass(frozen=True)
-class BlockReservationLink(ReservationLink):
-  block_name: str
-
-
-@dataclass(frozen=True)
-class SubBlockReservationLink(BlockReservationLink):
-  sub_block_name: str
-
-
-def print_reservations(args) -> int:
-  """Print the reservations in the project.
-
-  Args:
-    args: user provided arguments for running the command.
-
-  Returns:
-    0 if successful and 1 otherwise.
-  """
-  command = f'gcloud beta compute reservations list --project={args.project}'
-  return_code = run_command_with_updates(
-      command, 'Get all reservations in the project'
-  )
-  if return_code != 0:
-    xpk_print(f'Get all reservations returned ERROR {return_code}')
-    return 1
-  return 0
 
 
 def get_capacity_type(args) -> tuple[CapacityType, int]:
@@ -123,116 +89,6 @@ def get_capacity_type(args) -> tuple[CapacityType, int]:
     return_code = 1
 
   return capacity_type, return_code
-
-
-def get_reservation_maintenance_interval(reservation: ReservationLink) -> str:
-  """Get reservation maintenance interval.
-
-  Args:
-    reservation: reservation object.
-
-  Returns:
-    Maintenance interval as a string.
-  """
-  command = (
-      f'gcloud beta compute reservations describe {reservation.name}'
-      f' --project={reservation.project} --zone={reservation.zone} --format="value(specificReservation.instanceProperties.maintenanceInterval)"'
-  )
-  return_code, output = run_command_for_value(
-      command, 'Get reservation maintenance interval'
-  )
-  if return_code != 0:
-    xpk_print(f'Get reservation maintenance interval ERROR {return_code}')
-    xpk_exit(1)
-  return output.strip()
-
-
-def get_reservation_placement_policy(reservation: ReservationLink) -> str:
-  """Get reservation placement policy.
-
-  Args:
-    reservation: reservation object.
-
-  Returns:
-    Placement policy as a string.
-  """
-  command = (
-      f'gcloud beta compute reservations describe {reservation.name}'
-      f' --project={reservation.project} --zone={reservation.zone} --format="value(resourcePolicies.policy)"'
-  )
-  return_code, output = run_command_for_value(
-      command, 'Get reservation placement policy'
-  )
-  if return_code != 0:
-    xpk_print(f'Get reservation placement policy ERROR {return_code}')
-    xpk_exit(1)
-  return output.strip()
-
-
-def get_reservation_deployment_type(reservation: ReservationLink) -> str:
-  """Get reservation deployment type.
-
-  Args:
-    reservation: reservation object.
-
-  Returns:
-    Deployment type as a string.
-  """
-  command = (
-      f'gcloud beta compute reservations describe {reservation.name}'
-      f' --project={reservation.project} --zone={reservation.zone} --format="value(deploymentType)"'
-  )
-  return_code, output = run_command_for_value(
-      command, 'Get reservation deployment type', dry_run_return_val='DENSE'
-  )
-  if return_code != 0:
-    xpk_print(f'Get reservation deployment type ERROR {return_code}')
-    xpk_exit(1)
-  return output.strip()
-
-
-def get_reservations_list(args) -> list[ReservationLink]:
-  """Get the list of reservations from args.
-
-  Args:
-    args: user provided arguments.
-
-  Returns:
-    List of ReservationLink objects.
-  """
-  if not args.reservation:
-    return []
-  return [
-      parse_reservation(r.strip(), args.project, args.zone)
-      for r in args.reservation.split(',')
-  ]
-
-
-def verify_reservations_exist(args) -> int:
-  """Verify the reservations exist.
-
-  Args:
-    args: user provided arguments for running the command.
-
-  Returns:
-    0 if successful and 1 otherwise.
-  """
-  for reservation in get_reservations_list(args):
-    command = (
-        f'gcloud beta compute reservations describe {reservation.name}'
-        f' --project={reservation.project} --zone={reservation.zone}'
-    )
-    return_code = run_command_with_updates(
-        command, 'Describe reservation', verbose=False
-    )
-    if return_code != 0:
-      xpk_print(f'Describe reservation returned ERROR {return_code}')
-      xpk_print(
-          f'Please confirm that your reservation name {reservation.name} is'
-          ' correct.'
-      )
-      return 1
-  return 0
 
 
 def get_capacity_arguments_from_capacity_type(
@@ -319,74 +175,3 @@ def get_capacity_node_selectors_from_capacity_type(
       )
       return_code = 1
   return node_selector, return_code
-
-
-def parse_reservation(
-    reservation_path: str, cluster_project: str, zone: str
-) -> ReservationLink:
-  """Parses the reservation details from the reservation path.
-
-  Args:
-    reservation_path: path in format `[projects/RESERVATION_PROJECT_ID/reservations/]RESERVATION_NAME[/reservationBlocks/BLOCK_NAME[/reservationSubBlocks/SUB_BLOCK_NAME]]`
-    cluster_project: the default cluster project
-    zone: the reservation zone
-
-  Returns:
-    ReservationLink instance containing reservation details.
-  """
-  reservation = _try_parse_reservation(reservation_path, cluster_project, zone)
-  if reservation is None:
-    xpk_print('Unable to parse reservation: ', reservation_path)
-    xpk_exit(1)
-  return reservation
-
-
-def _try_parse_reservation(
-    reservation_path: str, cluster_project: str, zone: str
-) -> ReservationLink | None:
-  parts = reservation_path.split('/')
-  if not all(parts):
-    return None
-
-  project = cluster_project
-  if parts[0] == 'projects':
-    if len(parts) < 4 or parts[2] != 'reservations':
-      return None
-    project = parts[1]
-    parts = parts[3:]  # remove projects/PROJECT/reservations/ prefix
-
-  match len(parts):
-    case 1:
-      return ReservationLink(project=project, name=parts[0], zone=zone)
-    case 3 if parts[1] == 'reservationBlocks':
-      return BlockReservationLink(
-          project=project, name=parts[0], zone=zone, block_name=parts[2]
-      )
-    case 5 if (
-        parts[1] == 'reservationBlocks' and parts[3] == 'reservationSubBlocks'
-    ):
-      return SubBlockReservationLink(
-          project=project,
-          name=parts[0],
-          zone=zone,
-          block_name=parts[2],
-          sub_block_name=parts[4],
-      )
-    case _:
-      return None
-
-
-def to_reservation_path(
-    reservation: ReservationLink, cluster_project: str
-) -> str:
-  """Convert reservation to path string."""
-  if reservation.project == cluster_project:
-    path = reservation.name
-  else:
-    path = f'projects/{reservation.project}/reservations/{reservation.name}'
-
-  if isinstance(reservation, BlockReservationLink):
-    path += f'/reservationBlocks/{reservation.block_name}'
-    if isinstance(reservation, SubBlockReservationLink):
-      path += f'/reservationSubBlocks/{reservation.sub_block_name}'
-  return path
