+++ src/xpk/core/nodepool.py
@@ -25,6 +25,8 @@ from .capacity import (
     CapacityType,
     get_capacity_arguments_from_capacity_type,
     get_capacity_type,
+)
+from .reservation import (
     get_reservations_list,
     print_reservations,
     to_reservation_path,
