+++ src/xpk/core/nodepool_test.py
@@ -26,7 +26,7 @@ from xpk.core.nodepool import (
 from xpk.core.system_characteristics import AcceleratorType, SystemCharacteristics, DockerPlatform, GpuConfig
 from xpk.core.commands import FailedCommand
 from xpk.core.testing.commands_tester import CommandsTester
-from xpk.core.capacity import ReservationLink
+from xpk.core.reservation import ReservationLink
 
 
 CLUSTER_NAME = "running-cucumber"
