+++ src/xpk/core/reservation_test.py
@@ -1,0 +1,237 @@
+"""
+Copyright 2025 Google LLC
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+     https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+
+import pytest
+from unittest.mock import MagicMock, patch
+from .reservation import (
+    get_reservation_deployment_type,
+    parse_reservation,
+    ReservationLink,
+    BlockReservationLink,
+    SubBlockReservationLink,
+    verify_reservations_exist,
+    get_reservations_list,
+    to_reservation_path,
+)
+
+
+@patch('xpk.core.reservation.xpk_print')
+def test_get_reservation_deployment_type_exits_with_command_fails(
+    xpk_print: MagicMock, mocker
+):
+  mocker.patch(
+      target='xpk.core.reservation.run_command_for_value', return_value=(1, '')
+  )
+  with pytest.raises(SystemExit):
+    get_reservation_deployment_type(
+        reservation=ReservationLink(
+            project='project', name='reservation', zone='zone'
+        )
+    )
+
+  assert (
+      'Get reservation deployment type ERROR 1'
+      in xpk_print.mock_calls[0].args[0]
+  )
+
+
+def test_get_reservation_deployment_type_returns_deployment_type_when_command_succeeds(
+    mocker,
+):
+  mocker.patch(
+      target='xpk.core.reservation.run_command_for_value',
+      return_value=(0, 'DENSE'),
+  )
+  result = get_reservation_deployment_type(
+      reservation=ReservationLink(
+          project='project', name='reservation', zone='zone'
+      )
+  )
+  assert result == 'DENSE'
+
+
+@pytest.mark.parametrize(
+    argnames='reservation_path,expected_reservation',
+    argvalues=[
+        (
+            'reservation',
+            ReservationLink(
+                project='cluster-project', name='reservation', zone='zone'
+            ),
+        ),
+        (
+            'reservation/reservationBlocks/block',
+            BlockReservationLink(
+                project='cluster-project',
+                name='reservation',
+                zone='zone',
+                block_name='block',
+            ),
+        ),
+        (
+            'reservation/reservationBlocks/block/reservationSubBlocks/subblock',
+            SubBlockReservationLink(
+                project='cluster-project',
+                name='reservation',
+                zone='zone',
+                block_name='block',
+                sub_block_name='subblock',
+            ),
+        ),
+        (
+            'projects/project/reservations/reservation',
+            ReservationLink(project='project', name='reservation', zone='zone'),
+        ),
+        (
+            'projects/project/reservations/reservation/reservationBlocks/block',
+            BlockReservationLink(
+                project='project',
+                name='reservation',
+                zone='zone',
+                block_name='block',
+            ),
+        ),
+        (
+            'projects/project/reservations/reservation/reservationBlocks/block/reservationSubBlocks/subblock',
+            SubBlockReservationLink(
+                project='project',
+                name='reservation',
+                zone='zone',
+                block_name='block',
+                sub_block_name='subblock',
+            ),
+        ),
+    ],
+)
+def test_parse_reservation_parses_valid_reservations(
+    reservation_path: str,
+    expected_reservation: ReservationLink,
+):
+  actual_reservation = parse_reservation(
+      reservation_path, 'cluster-project', 'zone'
+  )
+
+  assert actual_reservation == expected_reservation
+
+
+def test_to_reservation_path_for_reservation():
+  res = ReservationLink(project='project', name='reservation', zone='zone')
+  assert to_reservation_path(res, 'project') == 'reservation'
+  assert (
+      to_reservation_path(res, 'other-project')
+      == 'projects/project/reservations/reservation'
+  )
+
+
+def test_to_reservation_path_for_block_reservation():
+  res_block = BlockReservationLink(
+      project='project', name='reservation', zone='zone', block_name='block'
+  )
+  assert (
+      to_reservation_path(res_block, 'project')
+      == 'reservation/reservationBlocks/block'
+  )
+  assert (
+      to_reservation_path(res_block, 'other-project')
+      == 'projects/project/reservations/reservation/reservationBlocks/block'
+  )
+
+
+def test_to_reservation_path_for_sub_block_reservation():
+  res_sub = SubBlockReservationLink(
+      project='project',
+      name='reservation',
+      zone='zone',
+      block_name='block',
+      sub_block_name='subblock',
+  )
+  assert (
+      to_reservation_path(res_sub, 'project')
+      == 'reservation/reservationBlocks/block/reservationSubBlocks/subblock'
+  )
+  assert (
+      to_reservation_path(res_sub, 'other-project')
+      == 'projects/project/reservations/reservation/reservationBlocks/block/reservationSubBlocks/subblock'
+  )
+
+
+@pytest.mark.parametrize(
+    argnames='reservation_path',
+    argvalues=[
+        '',
+        '/name',
+        'name/',
+        'name/reservationBlocks/',
+        'name/reservationBlocks/block/reservationSubBlocks/',
+        'name/reservationBlocks/block/reservationSubBlocks/subblock/extra',
+        'name/reservationBlock/block/reservationSubBlocks/subblock',
+        'name/reservationBlocks/block/reservationSubBlock/subblock',
+        'reservations/name',
+        'project/project/reservations/name',
+        'projects/project/reservation/name',
+        'projects/project/reservations',
+        'projects/project/reservations/name/reservationBlocks/block/reservationSubBlocks/subblock/extra',
+        'projects/project/reservations/name/reservationBlocks//reservationSubBlocks/subblock',
+    ],
+)
+@patch('xpk.core.reservation.xpk_print')
+def test_parse_reservation_fails_on_invalid_reservations(
+    xpk_print: MagicMock, reservation_path: str
+):
+  with pytest.raises(SystemExit):
+    parse_reservation(reservation_path, 'cluster-project', 'zone')
+
+  assert 'Unable to parse reservation' in xpk_print.mock_calls[0].args[0]
+
+
+@patch('xpk.core.reservation.run_command_with_updates', return_value=0)
+def test_verify_reservations_exist_multiple(mock_run, mocker):
+  args = mocker.Mock(reservation='res1,res2', project='project', zone='zone')
+  assert verify_reservations_exist(args) == 0
+  assert mock_run.call_count == 2
+
+
+def test_get_reservations_list_with_single_reservation(mocker):
+  args = mocker.Mock(reservation='res1', project='project', zone='zone')
+  assert get_reservations_list(args) == [
+      ReservationLink(project='project', name='res1', zone='zone')
+  ]
+
+
+def test_get_reservations_list_with_multiple_reservations(mocker):
+  args = mocker.Mock(reservation='res1,res2', project='project', zone='zone')
+  assert get_reservations_list(args) == [
+      ReservationLink(project='project', name='res1', zone='zone'),
+      ReservationLink(project='project', name='res2', zone='zone'),
+  ]
+
+
+def test_get_reservations_list_with_whitespace(mocker):
+  args = mocker.Mock(reservation='res1, res2 ', project='project', zone='zone')
+  assert get_reservations_list(args) == [
+      ReservationLink(project='project', name='res1', zone='zone'),
+      ReservationLink(project='project', name='res2', zone='zone'),
+  ]
+
+
+def test_get_reservations_list_none(mocker):
+  args = mocker.Mock(reservation=None)
+  assert get_reservations_list(args) == []
+
+
+def test_get_reservations_list_empty(mocker):
+  args = mocker.Mock(reservation='')
+  assert get_reservations_list(args) == []
